Refactor name package
