+++ src/main/java/seedu/address/logic/commands/EditModuleClassCommand.java
@@ -14,7 +14,7 @@ import seedu.address.commons.core.index.Index;
 import seedu.address.commons.util.CollectionUtil;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
-import seedu.address.model.components.Name;
+import seedu.address.model.components.name.Name;
 import seedu.address.model.moduleclass.ModuleClass;
 
 /**
