+++ src/main/java/seedu/address/logic/commands/EditStudentCommand.java
@@ -19,7 +19,7 @@ import seedu.address.commons.core.index.Index;
 import seedu.address.commons.util.CollectionUtil;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
-import seedu.address.model.components.Name;
+import seedu.address.model.components.name.Name;
 import seedu.address.model.student.Email;
 import seedu.address.model.student.Student;
 import seedu.address.model.student.Telegram;
