+++ src/main/java/seedu/address/logic/parser/AddModuleClassCommandParser.java
@@ -7,7 +7,7 @@ import java.util.stream.Stream;
 
 import seedu.address.logic.commands.AddModuleClassCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.components.Name;
+import seedu.address.model.components.name.Name;
 import seedu.address.model.moduleclass.ModuleClass;
 
 /**
