+++ src/main/java/seedu/address/model/student/Student.java
@@ -8,8 +8,8 @@ import java.util.Objects;
 import java.util.Set;
 import java.util.UUID;
 
-import seedu.address.model.components.HasName;
-import seedu.address.model.components.Name;
+import seedu.address.model.components.name.HasName;
+import seedu.address.model.components.name.Name;
 import seedu.address.model.tag.Tag;
 
 /**
