+++ src/main/java/seedu/address/model/util/SampleDataUtil.java
@@ -6,7 +6,7 @@ import java.util.stream.Collectors;
 
 import seedu.address.model.ReadOnlyTutorsPet;
 import seedu.address.model.TutorsPet;
-import seedu.address.model.components.Name;
+import seedu.address.model.components.name.Name;
 import seedu.address.model.student.Email;
 import seedu.address.model.student.Student;
 import seedu.address.model.student.Telegram;
