+++ src/test/java/seedu/address/logic/parser/EditStudentCommandParserTest.java
@@ -31,7 +31,7 @@ import org.junit.jupiter.api.Test;
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.EditStudentCommand;
 import seedu.address.logic.commands.EditStudentCommand.EditStudentDescriptor;
-import seedu.address.model.components.Name;
+import seedu.address.model.components.name.Name;
 import seedu.address.model.student.Email;
 import seedu.address.model.student.Telegram;
 import seedu.address.model.tag.Tag;
