+++ src/test/java/seedu/address/testutil/EditStudentDescriptorBuilder.java
@@ -6,7 +6,7 @@ import java.util.stream.Stream;
 
 import seedu.address.logic.commands.EditStudentCommand;
 import seedu.address.logic.commands.EditStudentCommand.EditStudentDescriptor;
-import seedu.address.model.components.Name;
+import seedu.address.model.components.name.Name;
 import seedu.address.model.student.Email;
 import seedu.address.model.student.Student;
 import seedu.address.model.student.Telegram;
